requirements.txt
fastapi
uvicorn[standard]
networkx
python-dotenv
google-generativeai
pydantic
shapely
